added debugging info

+++ build.sh
@@ -1,5 +1,10 @@
-rm -rf build
-mkdir build
-cd build
-cmake ..
-make
+if [ $1 = debug ]; then
+    cmake -S . -B build -DCMAKE_BUILD_TYPE=Debug
+    cmake --build build
+else
+    rm -rf build
+    mkdir build
+    cd build
+    cmake ..
+    make
+fi
